Update guide to fix menu item issue
